+++ test-cli-docs.sh
@@ -5,6 +5,11 @@ die() { set +v; echo "$red$*$reset" 1>&2 ; exit 1; }
 
 for TOOL in src/*.py; do
   TOOL=`basename $TOOL`
+
+  # Simple enough not to need usage docs.
+  [ "$TOOL" == 'generate_field_descriptions.py' ] && continue
+
+
   echo "Testing $TOOL docs..."
   [ -e src/$TOOL ] || die "src/$TOOL does not exist."
   diff \